chore: fixing privacy policy
